[FE] refactor: 페이지네이션 컴포넌트화
